+++ ydb/public/lib/ydb_cli/dump/dump_impl.cpp
@@ -26,7 +26,8 @@ TDumpResult TDumpClient::Dump(const TString& dbPath, const TString& fsPath, cons
             settings.AvoidCopy_,
             settings.SavePartialResult_,
             settings.PreservePoolKinds_,
-            settings.Ordered_);
+            settings.Ordered_,
+            settings.Compression_);
         return Result<TDumpResult>();
     } catch (NBackup::TYdbErrorException& e) {
         return TDumpResult(std::move(e.Status));